frontend/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pyyaml>=6.0.1
mlflow>=2.9.0
numpy>=1.24.0
albumentations>=1.3.1